Round down to 2 decimals, instead of normal rounding
This is fixing a bug that only happened in a few cases. Imagine this scenario:

Lines covered: 14389
Total lines: 32862

In percentage it's: 43.786135962509896
If you round that to 2 decimals, it becomes: 43.79

This is a higher number than the real threshold, because we had to round upwards, so the next time we run jest tests, they will fail, because tests don't meet the coverage threshold.

The right way is to round down to 2 decimal, for which the result would be: 43.78